requirements.txt
freezegun
requests
kivy
docutils
pygments
pypiwin32
kivy-deps.sdl2
kivy-deps.glew
kivy-deps.gstreamer
kivy-deps.angle
kivy-garden
KivyCalendar
pytz
pytest
requests-mock
marshmallow
marshmallow_sqlalchemy
sqlalchemy
python-dateutil
-e git+git@github.com:evelyn9191/gtfspy.git@master#egg=gtfspy
sqlalchemy_utils
